+++ spec/tests/PlayerEvents.spec.ts
@@ -1,16 +1,18 @@
-/// <reference path='../../src/ts/Conviva.d.ts'/>
-
 import { MockHelper, TestingPlayerAPI } from '../helper/MockHelper';
 import { ConvivaAnalytics } from '../../src/ts';
+import * as Conviva from '@convivainc/conviva-js-coresdk';
+
+jest.mock('@convivainc/conviva-js-coresdk', () => {
+  const { MockHelper } = jest.requireActual('../helper/MockHelper');
+  return MockHelper.createConvivaMock();
+});
 
 describe('player event tests', () => {
   let convivaAnalytics: ConvivaAnalytics;
   let playerMock: TestingPlayerAPI;
-  let convivaVideoAnalytics: Conviva.ConvivaVideoAnalytics;
+  let convivaVideoAnalytics: Conviva.VideoAnalytics;
 
   beforeEach(() => {
-    MockHelper.mockConviva();
-
     playerMock = MockHelper.getPlayerMock();
     convivaVideoAnalytics = Conviva.Analytics.buildVideoAnalytics();
 
@@ -242,7 +244,6 @@ describe('player event tests', () => {
       expect(convivaVideoAnalytics.reportAdBreakStarted).toHaveBeenCalledWith(
         Conviva.Constants.AdType.CLIENT_SIDE,
         Conviva.Constants.AdPlayer.SEPARATE,
-        Conviva.Constants.AdPosition.PREROLL,
       );
     });
 
@@ -253,7 +254,6 @@ describe('player event tests', () => {
       expect(convivaVideoAnalytics.reportAdBreakStarted).toHaveBeenCalledWith(
         Conviva.Constants.AdType.CLIENT_SIDE,
         Conviva.Constants.AdPlayer.SEPARATE,
-        Conviva.Constants.AdPosition.MIDROLL,
       );
     });