+++ src/ts/ConvivaAnalytics.ts
@@ -1,4 +1,5 @@
-import {
+import * as Conviva from '@convivainc/conviva-js-coresdk';
+import type {
   AdBreak,
   AdBreakEvent,
   AdEvent,
@@ -47,7 +48,7 @@ export interface ConvivaAnalyticsConfiguration {
    * user agent string parsing by the Conviva SDK. (default: WEB)
    * @deprecated Use `deviceMetadata.category` field
    */
-  deviceCategory?: Conviva.Constants.DeviceCategory;
+  deviceCategory?: Conviva.valueof<Conviva.ConvivaConstants['DeviceCategory']>;
 
   /**
    * Option to override the Conviva Device Metadata.
@@ -59,7 +60,7 @@ export interface ConvivaAnalyticsConfiguration {
      * user agent string parsing by the Conviva SDK.
      * (default: The same specified in config.deviceCategory)
      */
-    category?: Conviva.Constants.DeviceCategory;
+    category?: Conviva.valueof<Conviva.ConvivaConstants['DeviceCategory']>;
 
     /**
      * Option to override the Conviva Device Brand.
@@ -83,7 +84,7 @@ export interface ConvivaAnalyticsConfiguration {
      * Option to override the Conviva Device Type
      * (Default: Auto extract from User Agent string)
      */
-    type?: Conviva.Constants.DeviceType;
+    type?: Conviva.valueof<Conviva.ConvivaConstants['DeviceType']>;
 
     /**
      * Option to override the Conviva Device Version.
@@ -121,7 +122,7 @@ export class ConvivaAnalytics {
 
   private readonly logger: Conviva.LoggingInterface;
   private sessionKey: number;
-  private convivaVideoAnalytics: Conviva.ConvivaVideoAnalytics;
+  private convivaVideoAnalytics: Conviva.VideoAnalytics;
 
   /**
    * Tracks the ad playback status and is true between ON_AD_STARTED and ON_AD_FINISHED/SKIPPED/ERROR.
@@ -185,7 +186,7 @@ export class ConvivaAnalytics {
     this.isAd = false;
 
     const deviceMetadataFromConfig = this.config.deviceMetadata || {};
-    const deviceMetadata = {
+    const deviceMetadata: Conviva.ConvivaDeviceMetadata = {
       [Conviva.Constants.DeviceMetadata.CATEGORY]:
         deviceMetadataFromConfig.category || this.config.deviceCategory || Conviva.Constants.DeviceCategory.WEB,
       [Conviva.Constants.DeviceMetadata.BRAND]: deviceMetadataFromConfig.brand,
@@ -198,7 +199,7 @@ export class ConvivaAnalytics {
     };
     Conviva.Analytics.setDeviceMetadata(deviceMetadata);
 
-    let callbackFunctions = {} as { [key: string]: Function };
+    let callbackFunctions: Record<string, Function> = {};
     callbackFunctions[Conviva.Constants.CallbackFunctions.CONSOLE_LOG] = this.logger.consoleLog;
     callbackFunctions[Conviva.Constants.CallbackFunctions.MAKE_REQUEST] = new Html5Http().makeRequest;
     const html5Storage = new Html5Storage();
@@ -207,7 +208,7 @@ export class ConvivaAnalytics {
     callbackFunctions[Conviva.Constants.CallbackFunctions.CREATE_TIMER] = new Html5Timer().createTimer;
     callbackFunctions[Conviva.Constants.CallbackFunctions.GET_EPOCH_TIME_IN_MS] = new Html5Time().getEpochTimeMs;
 
-    const settings = {} as { [key: string]: string | number };
+    const settings: Record<string, string | number> = {};
     settings[Conviva.Constants.GATEWAY_URL] = config.gatewayUrl;
     settings[Conviva.Constants.LOG_LEVEL] = this.config.debugLoggingEnabled
       ? Conviva.Constants.LogLevel.DEBUG
@@ -318,7 +319,7 @@ export class ConvivaAnalytics {
    * @param metadataOverrides Metadata attributes which will be used to track to conviva.
    * @see ContentMetadataBuilder for more information about permitted attributes
    */
-  public updateContentMetadata(metadataOverrides: Metadata) {
+  public updateContentMetadata(metadataOverrides: Partial<Metadata>) {
     this.internalUpdateContentMetadata(metadataOverrides);
   }
 
@@ -332,7 +333,7 @@ export class ConvivaAnalytics {
    */
   public reportPlaybackDeficiency(
     message: string,
-    severity: Conviva.Constants.ErrorSeverity,
+    severity: Conviva.valueof<Conviva.ConvivaConstants['ErrorSeverity']>,
     endSession: boolean = true,
   ) {
     if (!this.isSessionActive()) {
@@ -354,7 +355,6 @@ export class ConvivaAnalytics {
     this.convivaVideoAnalytics.reportAdBreakStarted(
       Conviva.Constants.AdType.CLIENT_SIDE,
       Conviva.Constants.AdPlayer.SEPARATE,
-      Conviva.Constants.AdPosition.PREROLL,
     );
     this.debugLog('Tracking paused.');
   }
@@ -402,7 +402,7 @@ export class ConvivaAnalytics {
     }
   }
 
-  private internalUpdateContentMetadata(metadataOverrides: Metadata) {
+  private internalUpdateContentMetadata(metadataOverrides: Partial<Metadata>) {
     this.contentMetadataBuilder.setOverrides(metadataOverrides);
 
     if (!this.isSessionActive()) {
@@ -443,12 +443,19 @@ export class ConvivaAnalytics {
 
     // Create a Conviva monitoring session.
     this.convivaVideoAnalytics = Conviva.Analytics.buildVideoAnalytics();
+
+    this.convivaVideoAnalytics.setPlayerInfo({
+      [Conviva.Constants.FRAMEWORK_NAME]: 'Bitmovin Player',
+      [Conviva.Constants.FRAMEWORK_VERSION]: this.player.version,
+    });
+
     this.convivaVideoAnalytics.reportPlaybackRequested(this.contentMetadataBuilder.build());
     this.sessionKey = this.convivaVideoAnalytics.getSessionId();
     this.convivaVideoAnalytics.setCallback(() => {
       const playheadTimeMs = this.player.getCurrentTime('relativetime' as TimeMode) * 1000;
       this.convivaVideoAnalytics.reportPlaybackMetric(Conviva.Constants.Playback.PLAY_HEAD_TIME, playheadTimeMs);
     });
+
     this.debugLog('[ ConvivaAnalytics ] start session', this.sessionKey);
 
     if (!this.isSessionActive()) {
@@ -484,12 +491,12 @@ export class ConvivaAnalytics {
       ? Conviva.ContentMetadata.StreamType.LIVE
       : Conviva.ContentMetadata.StreamType.VOD;
 
-    this.contentMetadataBuilder.custom = {
+    this.contentMetadataBuilder.addToCustom({
       // Autoplay and preload are important options for the Video Startup Time so we track it as custom tags
       autoplay: PlayerConfigHelper.getAutoplayConfig(this.player) + '',
       preload: PlayerConfigHelper.getPreloadConfig(this.player) + '',
       integrationVersion: ConvivaAnalytics.VERSION,
-    };
+    });
 
     const source = this.player.getSource();
 
@@ -502,12 +509,11 @@ export class ConvivaAnalytics {
   private buildSourceRelatedMetadata(source: SourceConfig) {
     this.contentMetadataBuilder.assetName = this.getAssetNameFromSource(source);
     this.contentMetadataBuilder.viewerId = this.contentMetadataBuilder.viewerId;
-    this.contentMetadataBuilder.custom = {
+    this.contentMetadataBuilder.addToCustom({
       playerType: this.player.getPlayerType(),
       streamType: this.player.getStreamType(),
       vrContentType: source.vr && source.vr.contentType,
-      ...this.contentMetadataBuilder.custom,
-    };
+    });
 
     this.contentMetadataBuilder.streamUrl = this.getUrlFromSource(source);
   }
@@ -696,7 +702,6 @@ export class ConvivaAnalytics {
     this.convivaVideoAnalytics.reportAdBreakStarted(
       Conviva.Constants.AdType.CLIENT_SIDE,
       Conviva.Constants.AdPlayer.SEPARATE,
-      adPosition,
     );
   };