+++ src/ts/helper/AdBreakHelper.ts
@@ -1,7 +1,11 @@
 import { AdBreak, PlayerAPI } from 'bitmovin-player';
+import * as Conviva from '@convivainc/conviva-js-coresdk';
 
 export class AdBreakHelper {
-  public static mapAdPosition(adBreak: AdBreak, player: PlayerAPI): Conviva.Constants.AdPosition {
+  public static mapAdPosition(
+    adBreak: AdBreak,
+    player: PlayerAPI,
+  ): Conviva.valueof<Conviva.ConvivaConstants['AdPosition']> {
     if (adBreak.scheduleTime <= 0) {
       return Conviva.Constants.AdPosition.PREROLL;
     }