+++ src/ts/index.ts
@@ -1,2 +1,5 @@
+// Import to extend Conviva types.
+import './conviva/ConvivaExtension';
+
 export { ConvivaAnalytics, ConvivaAnalyticsConfiguration, EventAttributes } from './ConvivaAnalytics';
-export { Metadata, CustomContentMetadata } from './ContentMetadataBuilder';
+export { Metadata } from './ContentMetadataBuilder';